+++ src/EhrgoHealth.Web.UnitTests/Class1.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Xunit;
+﻿using Xunit;
 
 namespace EhrgoHealth.Web.UnitTests
 {
