+++ src/EhrgoHealth.Web/App_Start/WebApiConfig.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web.Http;
+﻿using System.Web.Http;
 
 namespace EhrgoHealth.Web
 {
