+++ src/EhrgoHealth/Class1.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace EhrgoHealth
+﻿namespace EhrgoHealth
 {
     public class Class1
     {
